x11rb: Moved key_to_keysym() to keycodes.rs
